Transform printEntity to a proper React component
